+++ src/components/dashboard/forms/items/NewUnitFormTargets.js
@@ -10,8 +10,10 @@ import { AdUnit } from 'adex-models'
 import Grid from '@material-ui/core/Grid'
 import Autocomplete from 'components/common/autocomplete'
 import classnames from 'classnames'
-import Slider from 'react-toolbox/lib/slider'
 import { items as ItemsConstants } from 'adex-constants'
+import Slider from '@material-ui/lab/Slider'
+import Typography from '@material-ui/core/Typography'
+
 
 const { ItemsTypes, Locations, TargetWeightLabels, Genders, TARGET_MIN_AGE, TARGET_MAX_AGE } = ItemsConstants
 
@@ -40,7 +42,7 @@ const AcGenders = autocompleteGenders()
 const ages = (() => {
     let ages = []
     for (var index = TARGET_MIN_AGE; index <= TARGET_MAX_AGE; index++) {
-        ages.push({ label: index + '' })
+        ages.push({ label: index + '', value: index + '' })
     }
 
     return ages
@@ -74,9 +76,11 @@ class NewUnitFormTargets extends Component {
             <Autocomplete
                 id='location-targets-select'
                 direction="auto"
-                multiple={true}
+                multiple
+                openOnClick
                 onChange={this.handleTargetChange.bind(this, target, null)}
                 label={this.props.t('TARGET_LOCATION')}
+                placeholder={this.props.t('TARGET_LOCATION_PLACEHOLDER')}
                 source={AcLocations}
                 value={target.value}
                 suggestionMatch='anywhere'
@@ -91,9 +95,11 @@ class NewUnitFormTargets extends Component {
             <Autocomplete
                 id='genders-targets-select'
                 direction="auto"
-                multiple={true}
+                multiple
+                openOnClick
                 onChange={this.handleTargetChange.bind(this, target, null)}
                 label={this.props.t('TARGET_GENDERS')}
+                placeholder={this.props.t('TARGET_GENDERS_PLACEHOLDER')}
                 source={AcGenders}
                 value={target.value}
                 suggestionMatch='anywhere'
@@ -117,9 +123,10 @@ class NewUnitFormTargets extends Component {
                         <Autocomplete
                             id='age-target-from-select'
                             direction="auto"
-                            multiple={false}
+                            openOnClick
                             onChange={this.handleTargetChange.bind(this, target, 'from')}
                             label={this.props.t('TARGET_AGE_FROM')}
+                            placeholder={this.props.t('TARGET_AGE_FROM_PLACEHOLDER')}
                             source={ages.slice(0, (value.to | 0) + 1)}
                             value={value.from + ''}
                             suggestionMatch='anywhere'
@@ -132,9 +139,10 @@ class NewUnitFormTargets extends Component {
                         <Autocomplete
                             id='age-target-to-select'
                             direction="auto"
-                            multiple={false}
+                            openOnClick
                             onChange={this.handleTargetChange.bind(this, target, 'to')}
                             label={this.props.t('TARGET_AGE_TO')}
+                            placeholder={this.props.t('TARGET_AGE_TO_PLACEHOLDER')}
                             source={ages.slice(value.from | 0)}
                             value={value.to + ''}
                             suggestionMatch='anywhere'
@@ -150,27 +158,36 @@ class NewUnitFormTargets extends Component {
     Targets = ({ meta, t }) => {
         return (
             <div>
-                <Grid container >
+                <Grid container spacing={16}>
                     <Grid
                         item
-                        className={theme.targetsHead}
-
-                        lg={7}>
-                        {t('TARGET')}
-                    </Grid>
-                    <Grid item lg={5}>
-                        {t('WEIGHT')}
+                        container xs={12}
+                        spacing={16}
+                        alignItems='flex-end'
+                    >
+                        <Grid item xs={12} md={7}>
+                            {t('TARGET')}
+                        </Grid>
+                        <Grid item xs={12} md={5}>
+                            {t('WEIGHT')}
+                        </Grid>
                     </Grid>
                     {
                         (meta.targets || []).map((target) => {
-                            // if (target.name !== 'location') return null
                             return (
-                                <Grid container spacing={16} key={target.name}>
+                                <Grid
+                                    item
+                                    container
+                                    xs={12}
+                                    key={target.name}
+                                    spacing={16}
+                                    alignItems='flex-end'
+                                >
                                     <Grid
                                         item
                                         key={target.name}
-                                        className={theme.targetRow}
-                                        lg={7}
+                                        xs={12}
+                                        md={7}
                                     >
                                         {(() => {
                                             switch (target.name) {
@@ -184,21 +201,22 @@ class NewUnitFormTargets extends Component {
                                             }
                                         })()}
                                     </Grid>
-                                    <Grid item lg={5} style={{ position: 'relative' }}>
-                                        <div className={classnames(theme.sliderWrapper)}>
-                                            <label className={classnames(theme.sliderLabel, theme.weightLabel)}>
-                                                {target.name}  weight:
+                                    <Grid item xs={12} md={5} >
+                                        <div
+                                            style={{ flex: '1' }}
+                                        >
+                                            <Typography noWrap id="page-size">
+                                                <label className={classnames(theme.sliderLabel, theme.weightLabel)}>
+                                                    {target.name}  Weight:
                                             <strong> {target.weight} </strong>
-                                                ({TargetWeightLabels[target.weight].label})
-                                        </label>
-                                            <Slider className={theme.weightSlider}
-                                                pinned
-                                                snaps
-                                                min={0}
-                                                max={4}
+                                                    ({TargetWeightLabels[target.weight].label})
+                                        </label></Typography>
+                                            <Slider
+                                                aria-labelledby={target.name + '-weight'}
+                                                min={0} max={4}
                                                 step={1}
                                                 value={target.weight}
-                                                onChange={this.handleTargetChange.bind(this, target, 'updateWeight')}
+                                                onChange={(ev, val) => this.handleTargetChange(target, 'updateWeight', val)}
                                             />
                                         </div>
                                     </Grid>
